+++ wikidict/lang/fr/ko_hangeul.py
